cleaning the code, removing console logs
